Add some humor.
The original wit supposedly goes like this:
    "There are only two hard things in Computer Science:
        cache invalidation and naming things."

Allegedly, this little aphorism was coined by a guy named **Phil Karlton**.

Wikipedia doesn't seem to have an article on him. Some very narrow web
search seems to indicate that he was one of the leaders of the X11 sample
server design and implementation.

There are a few variations floating around on the web. This one's mine.
And it's not funny at all.